#13 Refactor integrations
