+++ src/data/resolvers/mutations/notifications.js
@@ -9,8 +9,8 @@ export default {
    * @param {Boolean} object2.isAllowed - Shows whether notifications will be received or not
    * @param {Object} object3 - Middleware data
    * @param {Object} object3.user - The user making this action
-   * @return {Promise} returns Promise resolving a Notification document
-   * @throws {Error} throws error if user is not logged in
+   * @return {Promise} return Promise resolving a Notification document
+   * @throws {Error} throws Error('Login required') if user is not logged in
    */
   notificationsSaveConfig(root, doc, { user }) {
     if (!user) {
@@ -28,7 +28,7 @@ export default {
    * @param {Object} object3 - Middleware data
    * @param {Object} object3.user - The user making this action
    * @return {Promise}
-   * @throws {Error} throws error if user is not logged in
+   * @throws {Error} throws Error('Login required') if user is not logged in
    */
   notificationsMarkAsRead(root, { ids }, { user }) {
     if (!user) {
